+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/ExoPlayerEventListener.java
@@ -93,6 +93,11 @@ public abstract class ExoPlayerEventListener implements Player.Listener {
   }
 
   @Override
+  public void onRenderedFirstFrame() {
+    events.onRenderedFirstFrame();
+  }
+
+  @Override
   public void onTracksChanged(@NonNull Tracks tracks) {
     // Find the currently selected audio track and notify
     String selectedAudioTrackId = findSelectedAudioTrackId(tracks);