+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayerCallbacks.java
@@ -17,6 +17,8 @@ import androidx.annotation.Nullable;
  * <p>See {@link androidx.media3.common.Player.Listener} for details.
  */
 public interface VideoPlayerCallbacks {
+  void onRenderedFirstFrame();
+
   void onInitialized(int width, int height, long durationInMs, int rotationCorrectionInDegrees);
 
   void onPlaybackStateChanged(@NonNull PlatformPlaybackState state);