+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayerEventCallbacks.java
@@ -60,6 +60,11 @@ final class VideoPlayerEventCallbacks implements VideoPlayerCallbacks {
   }
 
   @Override
+  public void onRenderedFirstFrame() {
+    eventSink.success(new FirstFrameRenderedEvent(android.os.SystemClock.elapsedRealtime()));
+  }
+
+  @Override
   public void onIsPlayingStateUpdate(boolean isPlaying) {
     eventSink.success(new IsPlayingStateEvent(isPlaying));
   }