+++ packages/video_player/video_player_android/lib/src/android_video_player.dart
@@ -528,6 +528,8 @@ class _PlayerInstance {
 
   void _onStreamEvent(PlatformVideoEvent event) {
     switch (event) {
+      case FirstFrameRenderedEvent _:
+        _eventStreamController.add(VideoEvent(eventType: VideoEventType.firstFrameRendered));
       case InitializationEvent _:
         _eventStreamController.add(
           VideoEvent(